update theme picker
