+++ docs/app/components/content/IconsThemeSelect.vue
@@ -1,0 +1,15 @@
+<script setup lang="ts">
+const { icon, icons } = useTheme()
+</script>
+
+<template>
+  <UFormField label="Theme preset">
+    <USelect
+      v-model="icon"
+      color="neutral"
+      :icon="icons.find(i => i.value === icon)?.icon"
+      :items="icons"
+      class="w-50"
+    />
+  </UFormField>
+</template>
