+++ docs/content/docs/1.getting-started/6.integrations/1.icons/1.nuxt.md
@@ -126,4 +126,6 @@ Read more about this in the `@nuxt/icon` documentation.
 
 You can change the default icons used by components in your `app.config.ts`:
 
+:icons-theme-select
+
 :icons-theme
