+++ docs/content/docs/1.getting-started/6.integrations/1.icons/2.vue.md
@@ -59,4 +59,6 @@ slots:
 
 You can change the default icons used by Nuxt UI components in your `vite.config.ts`:
 
+:icons-theme-select
+
 :icons-theme
