+++ docs/content/docs/1.getting-started/6.integrations/2.fonts.md
@@ -18,14 +18,7 @@ Nuxt UI automatically registers the [`@nuxt/fonts`](https://github.com/nuxt/font
 
 To use a font in your Nuxt UI application, you can simply declare it in your CSS. It will be automatically loaded and optimized for you.
 
-```css [app/assets/css/main.css]
-@import "tailwindcss";
-@import "@nuxt/ui";
-
-@theme {
-  --font-sans: 'Public Sans', sans-serif;
-}
-```
+:font-theme
 
 ### Configuration
